add login button test
